Main Controller Update
Allow ability to add a `settings.json` file to be added to a modules root to store settings related to the module. The json will automatically be decoded into the controllers $settings property.
Added RenderView method, allows you to render a view without an explicit echo
Added LoadModel, same as the original model method but better named.
render and model methods are being kept for backwards compatibility